refactor(ui): split the approvals page and move it onto colour tokens
1027 -> 274 lines; cards/rows to ApprovalRows.tsx, shapes to types.ts.
